+++ Addon/ElvUI_mMediaTag/core/changelog/3.68.lua
@@ -1,0 +1,8 @@
+mMT.Changelog[368] = {
+	DATE = "03.01.2025",
+	NEW = { "[Portraits] New Textures Thin Diamond and Thin Circle" },
+	UPDATE = {
+		"[System]: DB function optimized.",
+		"[System]: Retail TOC.",
+	},
+}